refactor(hub-discussions): move parseMentionedUsers and parseDiscussionURI from index to separate fi
affects: @esri/hub-discussions
