+++ packages/discussions/test/utils/posts/parse-discussion-uri.test.ts
@@ -1,0 +1,61 @@
+import { IGroup } from "@esri/arcgis-rest-types";
+import { IDiscussionParams } from "../../../src/types";
+import { parseDiscussionURI } from "../../../src/utils/posts";
+import { MENTION_ATTRIBUTE } from "../../../src/utils/constants";
+
+describe("parseDiscussionURI", () => {
+  it("returns DiscussionParams for valid discussion uri", () => {
+    const discussion = "hub://dataset/1234_1/?id=1,2,3&attribute=foo";
+    const expected = {
+      source: "hub",
+      type: "dataset",
+      id: "1234",
+      layer: "1",
+      features: "1,2,3".split(","),
+      attribute: "foo",
+    } as IDiscussionParams;
+
+    expect(parseDiscussionURI(discussion)).toEqual(expected);
+  });
+
+  it("defaults layer to null", () => {
+    const discussion = "hub://dataset/1234/";
+    const expected = {
+      source: "hub",
+      type: "dataset",
+      id: "1234",
+      layer: null,
+      features: null,
+      attribute: null,
+    } as IDiscussionParams;
+
+    expect(parseDiscussionURI(discussion)).toEqual(expected);
+  });
+
+  it("returns DiscussionParams for invalid discussion uri", () => {
+    const discussion = "hub://dataset";
+    const expected = {
+      source: "hub",
+      type: "dataset",
+      id: null,
+      layer: null,
+      features: null,
+      attribute: null,
+    } as IDiscussionParams;
+
+    expect(parseDiscussionURI(discussion)).toEqual(expected);
+  });
+
+  it("throws when given invalid url string", () => {
+    const discussion = "hubdataset";
+
+    expect(() => parseDiscussionURI(discussion)).toThrowError();
+  });
+});
+
+describe("MENTION_ATTRIBUTE", () => {
+  it("should be defined", () => {
+    expect(MENTION_ATTRIBUTE).toBeDefined();
+    expect(MENTION_ATTRIBUTE).toEqual("data-mention");
+  });
+});
